+++ examples/deno/basic_ts/example.ts
@@ -1,12 +1,12 @@
 import { loadSync } from "https://deno.land/std@0.173.0/dotenv/mod.ts";
-import { config, EngineParameters, getJson } from "../../../mod.ts";
+import { config, getJson } from "../../../mod.ts";
 
 const { API_KEY: apiKey } = loadSync();
 const params = {
   engine: "google",
   q: "Coffee",
   api_key: apiKey,
-} satisfies EngineParameters<"google">;
+};
 
 // Show result as JSON (async/await)
 const response1 = await getJson(params);