+++ examples/node/pagination_ts_node_14_up/example.ts
@@ -9,7 +9,7 @@
  */
 
 import * as Dotenv from "dotenv";
-import { config, EngineParameters, getJson } from "serpapi";
+import { config, getJson } from "serpapi";
 
 Dotenv.config();
 const apiKey = process.env.API_KEY;
@@ -21,7 +21,7 @@ const params = {
   engine: "google",
   q: "Coffee",
   api_key: apiKey,
-} satisfies EngineParameters<"google">;
+};
 
 // Pagination (async/await)
 let page1 = await getJson(params);